requirements.txt
# AWS SDK
boto3==1.35.0
botocore==1.35.0

# Web framework (dashboard API)
flask==3.1.0
flask-cors==5.0.0

# ML & Data
prophet==1.1.6
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.4

# Database drivers
psycopg2-binary==2.9.10

# Environment management
python-dotenv==1.0.1

# Testing
pytest==8.3.5
pytest-html==4.1.1